CARLIN

Reachability algorithm based on conservative Carleman linearization procedure.

Kronecker product

ReachabilityAnalysis.kron_pow — Function

kron_pow(x::IA.Interval, pow::Int)

Given an interval x and an integer pow, compute x^pow.

Input

x – interval

pow – integer

Output

An interval enclosure of x^pow.

Examples

julia> [kron_pow(2 .. 3, i) for i in 1:3]
3-element Array{IntervalArithmetic.Interval{Float64},1}:
  [2, 3]
  [4, 9]
 [8, 27]

kron_pow(x::Interval, pow::Int)

Given an interval x and an integer pow, compute x^pow.

Input

x – interval

pow – integer

Output

An interval enclosure of x^pow as a LazySets Interval.

kron_pow(H::AbstractHyperrectangle, pow::Int)

Given hyperrectangular set H and an integer pow, compute the Kronecker power H^{⊗ pow}.

Input

H – hyperrectangular set

pow – integer power

Output

A hyperrectangle.

Algorithm

We compute H^{⊗ pow} where H is a hyperrectangular set by working with H as a product of intervals.

kron_pow(x::Vector{<:AbstractVariable}, pow::Int)

Compute the higher order concrete Kronecker power: x ⊗ x ⊗ ... ⊗ x, pow times for a vector of symbolic monomials.

Input

x – polynomial variable

pow – integer

Output

Vector of multivariate monomial corresponding to x^{⊗ pow}.

Examples

julia> using DynamicPolynomials

julia> @polyvar x[1:2]
(PolyVar{true}[x₁, x₂],)

julia> x
2-element Array{PolyVar{true},1}:
 x₁
 x₂

julia> kron_pow(x, 2)
4-element Array{Monomial{true},1}:
 x₁²
 x₁x₂
 x₁x₂
 x₂²

ReachabilityAnalysis.kron_pow_stack — Function

kron_pow_stack(x::Union{<:Interval, <:IA.Interval}, pow::Int)

Return an array with the interval powers [x, x^2, …, x^pow].

Input

x – interval

pow – integer power

Output

A vector of elements of the same type as x such that the i-th element is the interval x^i.

kron_pow_stack(H::AbstractHyperrectangle, pow::Int)

Return the Cartesian product array $H × H^{⊗2} × ⋯ × H^{⊗pow}$ where $H$ is a hyperrectangular set and $H^{⊗ i}$ is the i-th Kronecker power of $H$.

Input

H – hyperrectangular set

pow – integer power

Output

A Cartesian product array of hyperrectangles.

ReachabilityAnalysis.kron_id — Function

kron_id(n, k)

Compute the k-fold Kronecker product of the identity: I ⊗ I ⊗ ... ⊗ I, k times.

Input

n – integer representing the order (dimension of the identity)

k – integer representing the power

Output

A Diagonal matrix with element 1 in the diagonal and order $n^k$.

Examples

julia> kron_id(2, 2)
4×4 Diagonal{Float64,Array{Float64,1}}:
 1.0   ⋅    ⋅    ⋅
  ⋅   1.0   ⋅    ⋅
  ⋅    ⋅   1.0   ⋅
  ⋅    ⋅    ⋅   1.0

ReachabilityAnalysis.kron_sandwich — Function

kron_sandwich(F, n, k1, k2)

Compute A ⊗ F ⊗ C where A = I^{⊗ k1} and C = I^{⊗ k2} where I is the identity matrix of order n, the k1-fold and k2-fold Kronecker product of the identity and F in between.

Input

F – matrix

n – integer, dimension of the identity

k1 – nonnegative integer

k2 – nonnegative integer

Output

The kronecker product I^{⊗ k1} ⊗ F ⊗ I^{⊗ k2}, represented as a sparse matrix if F is sparse.

Examples

julia> F = sparse([0 1; -1 0.])
2×2 SparseMatrixCSC{Float64,Int64} with 2 stored entries:
  [2, 1]  =  -1.0
  [1, 2]  =  1.0

julia> Q = kron_sandwich(F, 2, 2, 2);

julia> size(Q)
(32, 32)

julia> I2 = I(2)
IdentityMultiple{Float64} of value 1.0 and order 2

julia> Q == reduce(kron, [I2, I2, F, I2, I2])
true

ReachabilityAnalysis.kron_sum — Function

kron_sum(F::AbstractMatrix, k::Int)

Compute the Kronecker sum of order k defined as:

\[ F ⊕ F ⊕ ... ⊕ F := F ⊗ I ⊗ ... ⊗ I + I ⊗ F ⊗ I ⊗ ... ⊗ I + ... + I ⊗ ... ⊗ I ⊗ F\]

where each term has k products and there are a total of k summands and I is the identity matrix of order n.

Examples

It holds that:

$k = 1: F$

$k = 2: F ⊗ I + I ⊗ F$

$k = 3: F ⊗ I ⊗ I + I ⊗ F ⊗ I + I ⊗ I ⊗ F$

julia> F = sparse([0 1; -1 0.])
2×2 SparseMatrixCSC{Float64,Int64} with 2 stored entries:
  [2, 1]  =  -1.0
  [1, 2]  =  1.0

julia> kron_sum(F, 1) == F
true

julia> I2 = I(2)
IdentityMultiple{Float64} of value 1.0 and order 2

julia> kron_sum(F, 2) == kron(F, I2) + kron(I2, F)
true

julia> kron_sum(F, 3) == sum(reduce(kron, X) for X in [[F, I2, I2], [I2, F, I2], [I2, I2, F]])
true

Error bounds

References