Internal functions and macros

This section of the manual includes some internal (i.e. unexported) functions and macros used within the library.

Internal functions and macros

Function annotations

Algorithm defaults

Function annotations

ReachabilityAnalysis.@requires — Macro

@requires(module_name)

Convenience macro to annotate that a package is required to use a certain function.

Input

module_name – name of the required package

Output

The macro expands to an assertion that checks whether the module module_name is known in the calling scope.

Notes

Usage:

function foo(...)
    @require MyPackage
    ... # functionality that requires MyPackage to be loaded
end

ReachabilityAnalysis.@commutative — Macro

@commutative(FUN)

Macro to declare that a given function FUN is commutative, returning the original FUN and a new method of FUN where the first and second arguments are swapped.

Input

FUN – function name

Output

A quoted expression containing the function definitions.

Algorithm defaults

ReachabilityAnalysis._default_cpost — Function

_default_cpost(ivp::IVP{<:AbstractContinuousSystem}, tspan; kwargs...)

Input

ivp – initial-value problem

tspan – time-span

Output

A continuous post-operator that can handle the given initial-value problem.

Notes

If the system is affine, then:

If it is one-dimensional, algorithm INT is used, otherwise,

Algorithm GLGM06 is used.

If the system is not affine, then the algorithm TMJets is used.