ReachabilityAnalysis.LGG09 — Type

LGG09{N, AM, TN<:AbstractDirections} <: AbstractContinuousPost

Implementation of Girard - Le Guernic algorithm for reachability analysis using support functions.

Fields

δ – step-size of the discretization

approx_model – (optional, default: Forward) approximation model; see Notes below for possible options

template – struct that holds the directions (either lazily or concretely) for each support function evaluation defining the template

static – (optional, default: false) if true, use statically sized arrays

threaded – (optional, default: true) if true, use multi-threading parallelism to compute the support function along each direction

sparse – (optional, default: false) if true, convert the matrix exponential obtained after discretization to a sparse matrix

cache – (optional, default: true) if true, use a cache for intermediate computations in the set recurrence loop

Notes

The type fields are:

N – number type of the step-size

AM – type of the approximation model

TN – type of the abstract directions that define the template

References

The is an implementation of the algorithm from [LGG09].

These methods are described at length in the dissertation [LG09].