Discretization

Discretization

Discretize API

Approximation models

Exponentiation

References

Discretize API

LinearAlgebra.normalize — Function

normalize(a::AbstractArray, p::Real=2)

Normalize the array a so that its p-norm equals unity, i.e. norm(a, p) == 1. See also normalize! and norm.

Examples

julia> a = [1,2,4];

julia> b = normalize(a)
3-element Array{Float64,1}:
 0.2182178902359924
 0.4364357804719848
 0.8728715609439696

julia> norm(b)
1.0

julia> c = normalize(a, 1)
3-element Array{Float64,1}:
 0.14285714285714285
 0.2857142857142857
 0.5714285714285714

julia> norm(c, 1)
1.0

julia> a = [1 2 4 ; 1 2 4]
2×3 Array{Int64,2}:
 1  2  4
 1  2  4

julia> norm(a)
6.48074069840786

julia> normalize(a)
2×3 Array{Float64,2}:
 0.154303  0.308607  0.617213
 0.154303  0.308607  0.617213

normalize(hs::HalfSpace{N}, p=N(2)) where {N}

Normalize a half-space.

Input

hs – half-space

p – (optional, default: 2) norm

Output

A new half-space whose normal direction $a$ is normalized, i.e., such that $‖a‖_p = 1$ holds.

normalize(P::AbstractHPolygon{N}, p=N(2)) where {N}

Normalize a polygon in constraint representation.

Input

P – polygon in constraint representation

p – (optional, default: 2) norm

Output

A new polygon in constraint representation whose normal directions $a_i$ are normalized, i.e., such that $‖a_i‖_p = 1$ holds.

normalize(P::HPoly{N}, p=N(2)) where {N}

Normalize a polyhedron in constraint representation.

Input

P – polyhedron in constraint representation

p – (optional, default: 2) norm

Output

A new polyhedron in constraint representation whose normal directions $a_i$ are normalized, i.e., such that $‖a_i‖_p = 1$ holds.

normalize(L::Line, p::Real=2.0)

Normalize the direction of a line.

Input

L – line

p – (optional, default: 2.0) vector p-norm used in the normalization

Output

A line whose direction has unit norm w.r.t the given p-norm.

Notes

See also normalize!(::Line, ::Real) for the in-place version.

normalize(system::AbstractSystem)

Transform a mathematical system to a normalized (or canonical) form.

Input

system – system; it can be discrete or continuous

Output

Either the same system if it already conforms to a canonical form, or a new system otherwise.

Notes

The normalization procedure consists of transforming a given system type into a "canonical" format that is used internally. More details are given below.

Algorithm

The implementation of normalize exploits MathematicalSystems's' types, which carry information about the problem as a type parameter.

Homogeneous ODEs of the form $x' = Ax, x ∈ \mathcal{X}$ are canonical if the associated problem is a ConstrainedLinearContinuousSystem and A is a matrix. This type does not handle non-deterministic inputs.

Note that a LinearContinuousSystem does not consider constraints on the state-space (such as an invariant); to specify state constraints, use a ConstrainedLinearContinuousSystem. If the passed system is a LinearContinuousSystem (i.e. no constraints) then the normalization fixes a universal set (Universe) as the constraint set.

The generalization to canonical systems with constraints and possibly time-varying non-deterministic inputs is considered next. These systems are of the form $x' = Ax + u, u ∈ \mathcal{U}, x ∈ \mathcal{X}$. The system type is ConstrainedLinearControlContinuousSystem, where A is a matrix, X is a set and U is an input, that is, any concrete subtype of AbstractInput.

If U is not given as an input, normalization accepts either a LazySet, or a vector of LazySets. In these cases, the sets are wrapped around an appropriate concrete input type.

If the system does not conform to a canonical form, the implementation tries to make the transformation; otherwise an error is thrown. In particular, ODEs of the form $x' = Ax + Bu$ are mapped into $x' = Ax + u, u ∈ B\mathcal{U}$, where now $u$ has the same dimensions as $x$.

The transformations described above are analogous in the discrete case, i.e. $x_{k+1} = A x_k$ and $x_{k+1} = Ax_{k} + u_k, u_k ∈ \mathcal{U}, x_k ∈ \mathcal{X}$ for the linear and affine cases respectively.

normalize(h::Histogram{T,N}; mode::Symbol=:pdf) where {T,N}

Normalize the histogram h.

Valid values for mode are:

:pdf: Normalize by sum of weights and bin sizes. Resulting histogram has norm 1 and represents a PDF.

:density: Normalize by bin sizes only. Resulting histogram represents count density of input and does not have norm 1. Will not modify the histogram if it already represents a density (h.isdensity == 1).

:probability: Normalize by sum of weights only. Resulting histogram represents the fraction of probability mass for each bin and does not have norm 1.

:none: Leaves histogram unchanged. Useful to simplify code that has to conditionally apply different modes of normalization.

Successive application of both :probability and :density normalization (in any order) is equivalent to :pdf normalization.

normalize(h::Histogram{T,N}, aux_weights::Array{T,N}...; mode::Symbol=:pdf) where {T,N}

Normalize the histogram h and rescales one or more auxiliary weight arrays at the same time (aux_weights may, e.g., contain estimated statistical uncertainties). The values of the auxiliary arrays are scaled by the same factor as the corresponding histogram weight values. Returns a tuple of the normalized histogram and scaled auxiliary weights.

ReachabilityAnalysis.discretize — Function

discretize(ivp::IVP, δ, alg::AbstractApproximationModel)

Set-based conservative discretization of a continuous-time initial value problem into a discrete-time problem.

Input

ivp – initial value problem for a linear ODE in canonical form (see Notes below)

δ – step size

alg – algorithm used to compute the approximation model, choose among:

julia> subtypes(ReachabilityAnalysis.AbstractApproximationModel)
5-element Vector{Any}:
 Backward
 CorrectionHull
 Forward
 NoBloating
 StepIntersect

Different approximation algorithms and their respective options are described in the docstring of each method, e.g. Forward.

Output

The initial value problem of a discrete system.

Notes

Initial-value problems considered in this function are of the form

\[x' = Ax(t) + u(t),\qquad x(0) ∈ \mathcal{X}_0,\qquad (1)\]

and where $u(t) ∈ U(k)$ add where $\{U(k)\}_k$ is a sequence of sets of non-deterministic inputs and $\mathcal{X}_0$ is the set of initial states. Recall that this initial-value problem is called homogeneous whenever U is the empty set. Other problems, e.g. $x' = Ax(t) + Bu(t)$ can be brought to the canonical form with the function normalize.

The initial value problem returned by this function consists of a set discretized (also called bloated) initial states $Ω₀$, together with the coefficient matrix $Φ = e^{Aδ}$ and a transformed sequence of inputs if $U$ is non-empty.

Two main variations of this algorithm are considered: dense time case and discrete time case.

In the dense time case, the transformation is such that the trajectories of the given continuous system are included in the computed flowpipe of the discretized system. More precisely, given a step size $δ$ and the system (1) conservative set-based discretization function computes a set, $Ω₀$, that guarantees to contain all the trajectories of (1) starting at any $x(0) ∈ \mathcal{X}_0$ and for any input function that satisfies $u(t) ∈ U(1)$, for any $t ∈ [0, δ]$. If $U$ is time-varying, this function also discretizes the inputs for $k ≥ 0$.

In the discrete time case, there is no bloating of the initial states and the input is assumed to remain constant between sampled times. Use the algorithm NoBloating() for this setting. If $U$ is time-varying, this function also discretizes the inputs for $k ≥ 0$.

There are algorithms to obatin such transformations, called approximation models in the technical literature. For references to the original papers, see the docstring of each concrete subtype of AbstractApproximationModel.

Approximation models

ReachabilityAnalysis.Forward — Type

Forward{EM, SO, SI, IT, BT} <: AbstractApproximationModel

Forward approximation model.

Fields

exp – exponentiation method

setops – set operations method

sih – symmetric interval hull

inv – (optional, default: false) if true, assume that the state matrix is invertible and use its inverse in the Φ functions

backend – (optional, default: nothing) used if the algorithm needs to apply concrete polyhedral computations

Algorithm

The transformations are:

$Φ ← \exp(Aδ)$,

$Ω_0 ← CH(\mathcal{X}_0, Φ\mathcal{X}_0 ⊕ δU(0) ⊕ E_ψ(U(0), δ) ⊕ E^+(\mathcal{X}_0, δ))$,

$V(k) ← δU(k) ⊕ E_ψ(U(k), δ)$.

Here we allow $U$ to be a sequence of time varying non-deterministic input sets.

For the definition of the sets $E_ψ$ and $E^+$ see [FRE11]. The Backward method uses $E^-$.

ReachabilityAnalysis.Backward — Type

Backward{EM, SO, SI, IT, BT} <: AbstractApproximationModel

Backward approximation model.

Fields

exp – exponentiation method

setops – set opertaions method

sih – symmetric interval hull

inv – (optional, default: false) if true, assume that the state matrix is invertible and use its inverse in the Φ functions

backend – (optional, default: nothing) used if the algorithm needs to apply concrete polyhedral computations

Algorithm

The transformations are:

$Φ ← \exp(Aδ)$,

$Ω_0 ← CH(\mathcal{X}_0, Φ\mathcal{X}_0 ⊕ δU(0) ⊕ E_ψ(U(0), δ) ⊕ E^-(\mathcal{X}_0, δ))$,

$V(k) ← δU(k) ⊕ E_ψ(U(k), δ)$.

Here we allow $U$ to be a sequence of time varying non-deterministic input sets.

For the definition of the sets $E_ψ$ and $E^-$ see [FRE11]. The Forward method uses $E^+$.

ReachabilityAnalysis.CorrectionHull — Type

CorrectionHull{EM} <: AbstractApproximationModel

Discretization using the correction hull of the matrix exponential.

Fields

exp – exponentiation method

order – order of the Taylor series expansion of the matrix exponential

Algorithm

For the homogeneous case, this method implements the transformation:

\[Ω_0 = CH(X_0, e^{Aδ} X_0) ⊕ FX_0\]

where $F$ is the correction (interval) matrix.

For the inhomogeneous case, $x' = Ax + u$, $x ∈ X, u ∈ U$, implements $Ω_0 = CH(X_0, exp(Aδ) X0) ⊕ FX0$ where $F$ is the correction (interval) matrix.

In both cases, if $A$ is an interval matrix, the exponential is overapproximated using methods from IntervalMatrices.jl.

ReachabilityAnalysis.NoBloating — Type

NoBloating{EM, SO, IT} <: AbstractApproximationModel

No bloating, or discrete-time, approximation model.

Fields

exp – exponentiation method

setops – set operations method

inv – (optional, default: false) if true, assume that the state matrix is invertible and use its inverse in the Φ functions

Algorithm

The transformations are:

$Φ ← \exp(Aδ)$

$Ω_0 ← \mathcal{X}_0$

$V(k) ← Φ₁(A, δ)U(k)$, $k ≥ 0$.

The function $Φ₁(A, δ)$ is defined in Φ₁. We allow $U$ to be a sequence of time varying non-deterministic input sets.

See also Eqs.(14) in [BFFPSV18].

Exponentiation

The state transition matrix of the linear ODE $x'(t) = Ax(t) + u(t)$ at time $\delta > 0$ is $\Phi = e^{A\delta}$, hence the algorithms usually require to compute exponential matrices. There are distinct ways to compute the matrix exponential $e^{A\delta}$ depending on the type of $A$ (see e.g. [HIH08]). The available methods can be used through the (unexported) function _exp.

For high dimensional systems (typicall n > 2000), computing the matrix exponential is expensive hence it is preferable to compute the action of the matrix exponential over vectors when needed, that is, $e^{δA} v$ for each $v$. This method is particularly well-suited if A is vert sparse. Use the option exp=:krylov (or exp=:lazy) for this purpose.

ReachabilityAnalysis._exp — Function

_exp(A::AbstractMatrix, δ, [alg]::AbstractExpAlg=BaseExp)

Compute the matrix exponential $e^{Aδ}$.

Input

A – matrix

δ – step size

alg – (optional, default: BaseExp) the algorithm used to take the matrix exponential of Aδ, possible options are BaseExp, LazyExp, PadeExp and IntervalExp see details in the Algorithm section below

Output

A matrix or a lazy wrapper of the matrix exponential, depending on alg.

Algorithm

BaseExp – (alias: :base) use Higham's scaling and squaring method implemented in Julia standard library; see ?exp for details; if A is a static array, uses the implementation in StaticArrays.jl

LazyExp – (alias: :lazy) return a lazy wrapper type around the matrix exponential using the implementation LazySets.SparseMatrixExp

PadeExp – (alias: pade) apply Pade approximant method to compute the matrix exponential of a sparse matrix (requires Expokit.jl)

IntervalExp – (alias: interval, taylor) apply the Taylor series expansion of the matrix exponential with an interval remainder; works if A is an interval matrix

Notes

If the algorithm LazyExp is used, evaluations of the action of the matrix exponential are done with the expmv implementation from Expokit (but see LazySets#1312 for the planned generalization to other backends).

ReachabilityAnalysis.Φ₁ — Function

Φ₁(A::AbstractMatrix, δ, [alg]::AbstractExpAlg=BaseExp, [isinv]::Bool=false, [Φ]=nothing)

Evaluate the series

\[Φ₁(A, δ) = ∑_{i=0}^∞ \dfrac{δ^{i+1}}{(i+1)!}A^i,\]

where $A$ is a square matrix of order $n$ and $δ > 0$.

Input

A – coefficients matrix

δ – step size

alg – (optional, default: BaseExp) the method used to take the matrix exponential of the coefficient matrix; see the documentation of _exp for available options

isinv – (optional, default: false) if true, assume that the coefficients matrix is invertible and compute $A^{-1}$

Φ – (optional, default: nothing) optionally pass the matrix exponential $e^{Aδ}$

Output

A matrix.

Algorithm

If $A$ is invertible, $Φ₁$ can be computed as

\[Φ₁(A, δ) = A^{-1}(e^{Aδ} - I_n),\]

where $I_n$ is the identity matrix of order $n$.

In the general case, implemented in this function, it can be computed as submatrices of the block matrix

\[P_{2n} = \exp \begin{pmatrix} Aδ && δI_n \\ 0 && 0 \end{pmatrix}.\]

It can be shown that

\[\exp(P_{2n}) = \begin{pmatrix} Φ(A, δ) && Φ₁(A, δ) \\ 0 && δI_n \end{pmatrix}.\]

where $Φ(A, δ) = e^{Aδ}$. In particular, Φ₁(A, δ) = P[1:n, (n+1):2*n]. This method can be found in [FRE11].

ReachabilityAnalysis.Φ₂ — Function

Φ₂(A::AbstractMatrix, δ, [alg]::AbstractExpAlg=BaseExp, [isinv]::Bool=false, [Φ]=nothing)

Evaluate the series

\[Φ₂(A, δ) = ∑_{i=0}^∞ \dfrac{δ^{i+2}}{(i+2)!}A^i,\]

where $A$ is a square matrix of order $n$ and $δ > 0$.

Input

A – coefficients smatrix

δ – step size

alg – (optional, default: BaseExp) the method used to take the matrix exponential of the coefficient matrix; see the documentation of _exp for available options

isinv – (optional, default: false) if true, assume that the coefficients matrix is invertible and compute $A^{-1}$

Φ – (optional, default: nothing) optionally pass the matrix exponential $e^{Aδ}$

Output

A matrix.

Algorithm

If $A$ is invertible, $Φ₂$ can be computed as

\[Φ₂(A, δ) = A^{-2}(e^{δA} - I_n - δA).\]

In the general case, implemented in this function, it can be computed as submatrices of the block matrix

\[P_{3n} = \exp \begin{pmatrix} Aδ && δI_n && 0 \\ 0 && 0 && δI_n \\ 0 && 0 && 0 \end{pmatrix}.\]

It can be shown that

\[\exp(P_{3n}) = \begin{pmatrix} Φ(A, δ) && Φ₁(A, δ) && 0 \\ 0 && 0 && δI_n \\ 0 && 0 && 0 \end{pmatrix}.\]

where $Φ(A, δ) = e^{Aδ}$. In particular, Φ₂ = P_{3n}[1:n, (2*n+1):3*n]. This method can be found in [FRE11].

References

HIH08Higham, Nicholas J. Functions of matrices: theory and computation. Society for Industrial and Applied Mathematics, 2008.