Discretization

Discretization

Approximation models

Discretize API

Exponentiation

Approximation models

ReachabilityAnalysis.AbstractApproximationModel — Type

AbstractApproximationModel

Abstract supertype for all approximation models.

Discretize API

TODO: document discretize.

Exponentiation

ReachabilityAnalysis._exp — Function

_exp(A::AbstractMatrix, δ::Float64, method::Symbol)

Compute the matrix exponential $e^{Aδ}$.

Input

A – matrix

δ – step size

method – symbol with the method used to take the matrix exponential of A; possible options are:

:base – use the scaling and squaring method implemented in Julia standard library; see ?exp for details

:lazy – return a lazy wrapper type around the matrix exponential using the implementation LazySets.SparseMatrixExp

:pade – apply Pade approximant method to compute the matrix exponential of a sparse matrix (requires Expokit)

Output

A matrix or a lazy wrapper of the matrix exponential, depending on method.

Notes

If the algorithm "lazy" is used, evaluations of the action of the matrix exponential are done with the expmv implementation from Expokit (but see LazySets#1312 for the planned generalization to other backends).

ReachabilityAnalysis.Φ₁ — Function

Φ₁(A, δ, method)

Compute the series

\[Φ₁(A, δ) = ∑_{i=0}^∞ \dfrac{δ^{i+1}}{(i+1)!}A^i,\]

where $A$ is a square matrix of order $n$ and $δ ∈ \mathbb{R}_{≥0}$.

Input

A – coefficient matrix

δ – step size

method – (optional, default: "base") the method used to take the matrix exponential of the coefficient matrix; see the documentation of _exp_Aδ for available options

Output

A matrix.

Algorithm

We use the method from [1]. If $A$ is invertible, $Φ₁$ can be computed as

\[Φ₁(A, δ) = A^{-1}(e^{δA} - I_n),\]

where $I_n$ is the identity matrix of order $n$.

In the general case, implemented in this function, it can be computed as submatrices of the block matrix

\[P = \exp \begin{pmatrix} Aδ && δI_n && 0 \\ 0 && 0 && δI_n \\ 0 && 0 && 0 \end{pmatrix}.\]

It can be shown that Φ₁(A, δ) = P[1:n, (n+1):2*n].

[1] Frehse, Goran, et al. "SpaceEx: Scalable verification of hybrid systems." International Conference on Computer Aided Verification. Springer, Berlin, Heidelberg, 2011.

ReachabilityAnalysis.Φ₂ — Function

Φ₂(A, δ, method)

Compute the series

\[Φ₂(A, δ) = ∑_{i=0}^∞ \dfrac{δ^{i+2}}{(i+2)!}A^i,\]

where $A$ is a square matrix of order $n$ and $δ ∈ \mathbb{R}_{≥0}$.

Input

A – coefficient matrix

δ – step size

method – the method used to take the matrix exponential of the coefficient matrix; see the documentation of _exp_Aδ for available options

Output

A matrix.

Algorithm

We use the method from [1]. If $A$ is invertible, $Φ₂$ can be computed as

\[Φ₂(A, δ) = A^{-2}(e^{δA} - I_n - δA).\]

In the general case, implemented in this function, it can be computed as submatrices of the block matrix

\[P = \exp \begin{pmatrix} Aδ && δI_n && 0 \\ 0 && 0 && δI_n \\ 0 && 0 && 0 \end{pmatrix}.\]

It can be shown that Φ₂(A, δ) = P[1:n, (2*n+1):3*n].

[1] Frehse, Goran, et al. "SpaceEx: Scalable verification of hybrid systems." International Conference on Computer Aided Verification. Springer, Berlin, Heidelberg, 2011.

Let $A ∈ \mathbb{R}^{n×n}$ and for $t ≥ 0$ consider the integral $\int_0^t e^{Aξ}dξ$. If $A$ is invertible, this integral (...)