Flowpipes

Abstract interface

ReachabilityAnalysis.AbstractFlowpipe — Type

AbstractFlowpipe

Abstract type representing a flowpipe.

Notes

A flowpipe is the set union of an array of reach-sets.

The following functions are available at the interface level.

ReachabilityAnalysis.basetype — Method

basetype(T::Type{<:AbstractFlowpipe})

Return the base type of the given flowpipe type (i.e., without type parameters).

Input

T – flowpipe type, used for dispatch

Output

The base type of T.

LazySets.ρ — Method

LazySets.ρ(d::AbstractVector, fp::AbstractFlowpipe)

Input

d – direction

fp – flowpipe

Output

The support function of the flowpipe along the given direction d.

Notes

In this fallback implementation, the flowpipe behaves like the union of the reach-sets, i.e. the implementation is analogue to that of a LazySet.UnionSetArray.

LazySets.σ — Method

LazySets.σ(d::AbstractVector, fp::AbstractFlowpipe)

Input

d – direction

fp – flowpipe

Output

The support vector of the flowpipe along the given direction d.

Notes

In this fallback implementation, the flowpipe behaves like the union of the reach-sets, i.e. the implementation is analogue to that of a LazySet.UnionSetArray.

LazySets.dim — Method

LazySets.dim(fp::AbstractFlowpipe)

Input

fp – flowpipe

Output

An integer representing the ambien dimension of the flowpipe.

ReachabilityAnalysis.set — Method

set(fp::AbstractFlowpipe, ind::Integer)

Return the geometric set represented by this flowpipe at the given index.

Input

fp – flowpipe

ind – index (from 1 to length(flowpipe))

Output

The set wrapped by the flowpipe at the given index.

ReachabilityAnalysis.tstart — Method

tstart(fp::AbstractFlowpipe)

Return the initial time of this flowpipe.

Input

fp – flowpipe

Output

A float representing the initial time of the given flowpipe. The fallback is computed by taking the initial time of the first reach-set.

ReachabilityAnalysis.tend — Method

tend(fp::AbstractFlowpipe)

Return the final time of this flowpipe.

Input

R – reach-set

Output

A float representing the initial time of the given flowpipe. The fallback is computed by taking the final time of the last reach-set.

ReachabilityAnalysis.tspan — Method

tspan(fp::AbstractFlowpipe)

Return time span of this flowpipe.

Input

fp – flowpipe

Output

The interval representing the time span of the given flowpipe. The fallback is computed as (tstart(fp), tend(fp)), see tstart(::AbstractFlowpipe) and tend(::AbstractFlowpipe) for details.

Flowpipe

ReachabilityAnalysis.Flowpipe — Type

Flowpipe{N, RT<:AbstractReachSet{N}} <: AbstractFlowpipe

Type that wraps a flowpipe.

Fields

Xk – set

ext – extension dictionary; field used by extensions

Notes

The dimension of the flowpipe corresponds to the dimension of the underlying reach-sets; in this type, it is is assumed that the dimension is the same for the different reach-sets.

The following methods are available.

ReachabilityAnalysis.shift — Function

shift(R::AbstractReachSet, t0::Number)

Perform a time-shift of the given reach-set.

Input

R – reach-set

t0 – number that corresponds to the time-shift

Output

A new reach-set of the same type of R such that its time-span has been shifted by t0.

shift(fp::Flowpipe{N, ReachSet{N, ST}}, t0::Number) where {N, ST}

Return the time-shifted flowpipe by the given number.

Input

fp – flowpipe

t0 – time shift

Output

A new flowpipe such that the time-span of each constituent reach-set has been shifted by t0.

Notes

See also Shift for the lazy counterpart.

Hybrid flowpipe

ReachabilityAnalysis.HybridFlowpipe — Type

HybridFlowpipe{N, D, FT<:AbstractFlowpipe, VOA<:VectorOfArray{N, D, Vector{FT}}} <: AbstractFlowpipe

Type that wraps a vector of flowpipes of the same type, such that they are contiguous in time.

Fields

Fk – vector of flowpipes

ext – (optional, default: empty) dictionary for extensions

Notes

This type assumes that the flowpipes are contiguous in time. This means that the final time the i-th flowpipe matches the start time of the i+1-th flowpipe.

Mixed flowpipe

ReachabilityAnalysis.MixedFlowpipe — Type

MixedFlowpipe{N, D, FT<:AbstractFlowpipe, VOA<:VectorOfArray{N, D, Vector{FT}}} <: AbstractFlowpipe

Type that wraps a vector of flowpipes of the same time, such that they are not necessarily contiguous in time.

Fields

Fk – vector of flowpipes

ext – (optional, default: empty) dictionary for extensions

Notes

This type does not assume that the flowpipes are contiguous in time.

Shifted flowpipe

ReachabilityAnalysis.ShiftedFlowpipe — Type

ShiftedFlowpipe{FT<:AbstractFlowpipe, NT<:Number} <: AbstractFlowpipe

Type that lazily represents a flowpipe that has been shifted in time.

Fields

F – original flowpipe

t0 – time shift

Notes

This type can wrap any concrete subtype of AbstractFlowpipe, and the extra field t0 is such that the time spans of each reach-set in F are shifted by the amount t0 (which should be a subtype of Number).

A convenience constructor alias Shift is given.

Partitioned flowpipe

ReachabilityAnalysis.PartitionedFlowpipe — Type

PartitionedFlowpipe{N, D, FT<:AbstractFlowpipe, VOA<:VectorOfArray{N, D, Vector{FT}}} <: AbstractFlowpipe

Type that wraps a vector of flowpipes of possibly different types.

Fields

Fk – vector of flowpipes

ext – (optional, default: empty) dictionary for extensions

Notes