Flowpipes

Abstract types

ReachabilityAnalysis.AbstractFlowpipe — Type

AbstractFlowpipe

Abstract type representing a flowpipe.

Notes

A flowpipe is the set union of an array of reach-sets.

ReachabilityAnalysis.AbstractReachSet — Type

AbstractReachSet{N}

Abstract type for all reach-sets types.

Notes

A reach-set is a set representation X associated to a given time span Δt.

In its simplest form, we represent reach-sets with a struct that wraps the tuple (X, Δt), where X corresponds to a geometric set, eg. a polytope, and Δt is the interval with the time span associated to this reach-set.

This type is parametric in N. The parameter N represents for the type of numerical coefficient used in the representation (typically, N = Float64).

Although concrete subtypes of AbstractReachSet may represent the set X in different ways, or carry additional information as an extra type field, they should all implement the AbstractReachSet interface to enable shared functionality for reach-set types. In particular, each concrete subtype should implement the following methods:

set – return the geometric set

setrep – return the type of the set representation

tspan – return the time interval span

tstart – return the initial time

tend – return the final time

dim – return ambient dimension of the reach-set

Interface functions

ReachabilityAnalysis.set — Function

set(R::AbstractReachSet)

Return the geometric set represented by this reach-set.

Input

R – reach-set

Output

The set wrapped by the given reach-set.

ReachabilityAnalysis.setrep — Function

setrep(R::AbstractReachSet)

Return the type of the set representation of this reach-set.

Input

R – reach-set

Output

Type of the set representation of the given reach-set.

ReachabilityAnalysis.tstart — Function

tstart(R::AbstractReachSet)

Return the initial time of this reach-set.

Input

R – reach-set

Output

A float representing the initial time of the given reach-set.

ReachabilityAnalysis.tend — Function

tend(R::AbstractReachSet)

Return the final time of this reach-set.

Input

R – reach-set

Output

A float representing the final time of the given reach-set.

ReachabilityAnalysis.tspan — Function

tspan(R::AbstractReachSet)

Return time span of this reach-set.

Input

R – reach-set

Output

The interval representing the time span of the given reach-set.

Reachable set

ReachabilityAnalysis.ReachSet — Type

ReachSet{N, ST<:LazySet{N}} <: AbstractLazyReachSet{N}

Type that wraps a reach-set using a LazySet as underlying representation.

Fields

X – set

Δt – time interval

Notes

A ReachSet is a struct representing (an approximation of) the reachable states for a given time interval. The type of the representation is ST, which may be any subtype LazySet. For efficiency reasons, ST should be concretely typed.

By assumption the coordinates in this reach-set are associated to the integers 1, …, n. The function vars_idx returns such tuple.

Sparse reachable set

ReachabilityAnalysis.SparseReachSet — Type

SparseReachSet{N, ST<:LazySet{N}, D} <: AbstractReachSet{N}

Type that wraps a reach-set using a LazySet as underlying representation, together with a tuple of variables associated to this reach-set.

Fields

X – set

Δt – time interval

vars_idx – tuple of variable indices represented by the set X

Notes

A SparseReachSet is a struct representing (an approximation of) the reachable states for a given time interval. The type of the representation is ST, which may be any subtype of LazySet (ideally, concrete). Moreover, this type also stores information about the variables (also named coordinates, or by abuse of notation, dimensions) corresponding to the set X.

For instance in the ambient space n=5, one may have a SparseReachSet whose variables tuple is vars = (4, 5, 6), i.e. representing a three-dimensional projection of the full-dimensional reach-set. In consequence, the dimension of X doesn't match the length of vars, in general

In this type, the parameter N represents the numerical type of the LazySet (typically, Float64), the type ST represents the set representation used, and D denotes the dimension of this sparse reach set. Note that, in contrast to ReachSet, for SparseReachSet the number of dimensions is part of the type information.

Type-stable flowpipe

ReachabilityAnalysis.Flowpipe — Type

Flowpipe{N, RT<:AbstractReachSet{N}} <: AbstractFlowpipe

Type that wraps a flowpipe.

Fields

Xk – set

ext – extension dictionary; field used by extensions

Notes

The dimension of the flowpipe corresponds to the dimension of the underlying reach-sets; in this type, it is is assumed that the dimension is the same for the different reach-sets.

Hybrid flowpipe

ReachabilityAnalysis.HybridFlowpipe — Type

HybridFlowpipe{N, D, FT<:AbstractFlowpipe, VOA<:VectorOfArray{N, D, Vector{FT}}} <: AbstractFlowpipe

Type that wraps a vector of flowpipes of possibly differen types.

Fields

Fk – vector of flowpipes

ext – (optional, default: empty) dictionary for extensions

Notes