Solutions

Abstract interface

ReachabilityAnalysis.AbstractSolution — Type

AbstractSolution

Abstract supertype of all solution types of a rechability problem.

Solution of a reachability problem

ReachabilityAnalysis.ReachSolution — Type

ReachSolution{FT<:AbstractFlowpipe, ST<:AbstractPost} <: AbstractSolution

Type that wraps the solution of a reachability problem as a sequence of lazy sets, and a dictionary of options.

Fields

Xk – the list of AbstractReachSets

alg – algorihm used

options – the dictionary of options

TODO: document other methods in solutions.jl.